backend/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
lxml==4.5.0
psycopg2==2.8.4
pytz==2019.3
soupsieve==2.0
sqlparse==0.3.0
termcolor==1.1.0
